can you introduce menu to switc between ocalhost:5173/datasets/bb35dc60-9274-4f93-be69-7352238b1430/groups and componets ocalhost:5173/datasets/bb35dc60-9274-4f93-be69-7352238b1430/components
